Make threadMPI compliant with MPI 3
- A lot of `const`s added to function signatures.
- Change the type of string size pointer in `MPI_Error_string`.

Reference: https://www.mpi-forum.org/docs/mpi-3.0/mpi30-report.pdf

We cast away `const` in `tMPI_Reduce_fast` and `tMPI_Scan`
since we don't have an "atomic const pointer" type.
The buffer is not modified, so we uphold the promise of `const`.

Refs #4933

+++ src/external/thread_mpi/src/errhandler.cpp
@@ -123,7 +123,7 @@ int tMPI_Error(tMPI_Comm comm, int tmpi_errno)
 }
 
 
-int tMPI_Error_string(int errorcode, char *strn, size_t *resultlen)
+int tMPI_Error_string(int errorcode, char *strn, int *resultlen)
 {
     if (errorcode < 0 || errorcode >= N_TMPI_ERR)
     {
@@ -212,7 +212,7 @@ int tMPI_Comm_get_errhandler(tMPI_Comm comm, tMPI_Errhandler *errhandler)
 void tmpi_errors_are_fatal_fn(tMPI_Comm *comm, int *err)
 {
     char   errstr[TMPI_MAX_ERROR_STRING];
-    size_t len;
+    int len;
 
     tMPI_Error_string(*err, errstr, &len);
     if (comm)
@@ -230,7 +230,7 @@ void tmpi_errors_are_fatal_fn(tMPI_Comm *comm, int *err)
 void tmpi_errors_return_fn(tMPI_Comm *comm, int *err)
 {
     char   errstr[TMPI_MAX_ERROR_STRING];
-    size_t len;
+    int len;
 
     tMPI_Error_string(*err, errstr, &len);
     if (comm)
